+++ test/ja_serializer/content_type_negotiation_test.exs
@@ -24,6 +24,32 @@ defmodule JaSerializer.ContentTypeNegotiationTest do
     assert result.resp_body == "success"
   end
 
+  test "Passes proper content-type and missing accept headers through" do
+    conn = Plug.Test.conn("POST", "/", %{})
+            |> put_req_header("content-type", @valid)
+    result = ExamplePlug.call(conn, [])
+    assert result.status == 200
+    assert result.resp_body == "success"
+  end
+
+  test "Passes proper content-type and */* accept headers through" do
+    conn = Plug.Test.conn("POST", "/", %{})
+            |> put_req_header("content-type", @valid)
+            |> put_req_header("accept", "*/*")
+    result = ExamplePlug.call(conn, [])
+    assert result.status == 200
+    assert result.resp_body == "success"
+  end
+
+  test "Passes proper content-type and application/* accept headers through" do
+    conn = Plug.Test.conn("POST", "/", %{})
+            |> put_req_header("content-type", @valid)
+            |> put_req_header("accept", "application/*")
+    result = ExamplePlug.call(conn, [])
+    assert result.status == 200
+    assert result.resp_body == "success"
+  end
+
   test "Passes no content-type and valid accept header through on DELETE" do
     conn = Plug.Test.conn("DELETE", "/", %{})
             |> put_req_header("accept", @valid)